Enable listen to be called correctly with socket fds.
Currently when a socket file-descriptor that has been previously
bind()ed is passed in to a listen() call the framework treats the
file-descriptor as a named pipe, rather than a TCP socket.

This consequence of this is that it is not possible to use socket
specific APIs such as remoteAddress.

This fix correctly determines when a file-descriptor is a socket
and attaches this to a TCP object, rather than a Pipe object.